Collecting Articles on Trump's 100 Days
This project is to collect a serries of news articles and headlines on
President Donald J. Trump's first 100 days in office, and to assemble
that data into a csv file that can then be transformed into a timeline
using Timeline JS. Currently, only articles from the NYT are being
shown, with plans to expand to other news organizations.

This project was made in preparation for the 2018 AggieMUN
"Crossroads:The Future of the Republican Party" committee.